+++ src/main/java/cn/phpst/mall/repository/SpuRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpuRepository extends JpaRepository<Spu,Long>  {
-    Spu findOneById(Long id);
+public interface SpuRepository extends JpaRepository<Spu,Integer>  {
+    Spu findOneById(Integer id);
 }
